Extract LoaderSpy to separate file
